Refactor unused lexers
