+++ src/lib.rs
@@ -1,8 +1,6 @@
 pub mod compiler;
 pub mod errors;
 pub mod lexer;
-pub mod lexer2;
-pub mod lexer3;
 pub mod naming;
 pub mod parser;
 pub mod protocol;
