+++ src/main.rs
@@ -1,13 +1,9 @@
 use std::process::exit;
 
-use logos::Logos;
 use new_nu_parser::compiler::Compiler;
-use new_nu_parser::lexer::Lexer;
-use new_nu_parser::lexer2::Lexer2;
-use new_nu_parser::lexer3::{lex, print_tokens, Token3};
+use new_nu_parser::lexer::{lex, print_tokens};
 use new_nu_parser::parser::Parser;
 use new_nu_parser::resolver::Resolver;
-use new_nu_parser::token::TokenType2;
 use new_nu_parser::typechecker::Typechecker;
 
 fn main() {
@@ -31,57 +27,6 @@ fn main() {
             eprintln!("can't find {}", fname);
             exit(1);
         };
-
-        // Lex
-
-        // let mut contents2 = contents.clone();
-        // contents2.push(0); // null-terminated string
-        // let mut lexer2 = Lexer2::new(&contents2);
-        // let mut token = lexer2.next();
-        // println!("t2: {:?}", token);
-        // while !matches!(token.token_type, TokenType2::Eof | TokenType2::Invalid) {
-        //     token = lexer2.next();
-        //     // println!("t2: {:?}", token)
-        // }
-
-        // let contents3 = contents.clone();
-        // let contents3_str = String::from_utf8(contents.clone()).expect("invalid utf8");
-        // let mut lexer3 = Token3::lexer(&contents3).spanned();
-        // let mut tokens = Vec::with_capacity(contents3.len());
-        // let mut i = 0;
-        // println!("Tokens3");
-        // while let Some((token, span)) = lexer3.next() {
-        //     let Ok(tok) = token else {
-        //         println!("Token3 Error: {:?} {:?}", token, span);
-        //         break;
-        //     };
-
-        //     println!(
-        //         "Token3 {i:4}: {:25} span: {:4} .. {:4} '{}'",
-        //         format!("{:?}", tok),
-        //         span.start,
-        //         span.end,
-        //         String::from_utf8_lossy(
-        //             contents3
-        //                 .get(span.start..span.end)
-        //                 .expect("lexer: missing source for span"),
-        //         )
-        //         .replace("\r", "\\r")
-        //         .replace("\n", "\\n")
-        //         .replace("\t", "\\t")
-        //     );
-        //     i += 1;
-        //     tokens.push(tok);
-        // }
-
-        // println!("Done");
-        // break;
-
-        // let mut lexer = Lexer::new(&contents, 0);
-        // lexer.lex();
-        // if do_print {
-        //     lexer.print();
-        // }
 
         let span_offset = compiler.span_offset();
         compiler.add_file(&fname, &contents);
